+++ app/src/main/java/com/mycodeflow/academyproject/FragmentMoviesList.kt
@@ -10,6 +10,8 @@ import com.mycodeflow.item.decorators.MovieListItemDecorator
 import com.mycodeflow.moviesadapters.MainMenuMovieListAdapter
 import androidx.lifecycle.ViewModelProviders
 import com.mycodeflow.data.MovieListModel
+import com.mycodeflow.work.CacheUpdateWorkManager
+import kotlinx.coroutines.launch
 import com.mycodeflow.viewmodels.MovieListViewModel as MovieListViewModel
 
 class FragmentMoviesList : BaseFragment() {
@@ -23,6 +25,8 @@ class FragmentMoviesList : BaseFragment() {
         if (context is MovieDetailsListener){
             clickListener = context
         }
+        //starting periodic background cache update work with WorkManager
+        startBgCacheUpdate(context)
     }
 
     override fun onCreateView(
@@ -50,6 +54,11 @@ class FragmentMoviesList : BaseFragment() {
         clickListener = null
         movieListAdapter = null
         rvMovieList = null
+    }
+
+    private fun startBgCacheUpdate(context: Context){
+        val worker = CacheUpdateWorkManager(context)
+        worker.startBackgroundWork()
     }
 
     private fun setupMoviesAdapter(){
